+++ know-your-votes/src/router/index.js
@@ -19,14 +19,6 @@ Vue.use(VueRouter)
     component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
   },
   {
-    path: '/map',
-    name: 'Map',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/Map.vue')
-  },
-  {
     path: '*',
     name: 'Invalid',
     component: Home,