+++ know-your-votes/src/views/About.vue
@@ -1,3 +1,4 @@
+// USE AS REFERENCE ONLY! Will not be in Prod
 <template>
   <div class="about">
     <h1>This is an about page</h1>